requirements.txt
click>=8.0.0
gitpython>=3.1.0
pandas>=1.5.0
numpy>=1.21.0
tqdm>=4.64.0
json5>=0.9.0
pyyaml>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
jupyter>=1.0.0

